- created spec file - moved objects file to separate folder
