+++ cypress/support/pageobjects/SequenceAnayzerApplication/sequence-analyzer.js
